Updates to bubble caller / ctx_place output
1. vcf_parse StrBuf* replaced with StrBuf to reduce de-referencing
2. bubble_caller takes max allele length and max flank length as params
3. bubble_caller print cmdline args in bubble header
4. traversal.c: updated, cmdline options added
5. Add traversal31 command to benchmarking
6. ctx_infer_edges.c: can merge colours when loading to save memory
7. db_graph.c: added add_all_edges
8. ctx_place.c: fixed naming bug: only filtered variants were given unique names

+++ src/common/file_reader.h
@@ -43,7 +43,7 @@ typedef struct
 } SeqLoadingStats;
 
 // Functions for dealing with file loading statistics
-SeqLoadingStats* seq_loading_stats_create(unsigned long readlen_arrsize);
+SeqLoadingStats* seq_loading_stats_create(size_t readlen_arrsize);
 void seq_loading_stats_sum(SeqLoadingStats* dst, SeqLoadingStats* src);
 void seq_loading_stats_free(SeqLoadingStats* stats);
